requirements.txt
Django==4.2.17
django-cors-headers==4.5.0
djangorestframework==3.15.2
fuzzywuzzy==0.18.0
gunicorn==23.0.0
openpyxl==3.1.5
pandas==2.1.2
python-multipart==0.0.20
RapidFuzz==3.13.0